Fix bug with JsonDecodeError catch
`requests` uses either simplejson or json depending on what json library
is available. catching `ValueError` allows to catch all cases. Here the
problem is that `kg-core-python` does not depend on `simplejson` so the
`JSONDecodeError` imported comes from the `json` package. On the other
hand, in the data-proxy, I have the `simplejson` dependency. In that
case `requests` uses it, and throws `simplejson.errors.JSONDecodeError`
instead. Thereforee the exception is not caught. One other solution to
the `ValueError` catch would be to have conditional import in
`kg-core-python`.